+++ etherparse/src/err/value_too_big_error.rs
@@ -1,0 +1,101 @@
+use crate::err;
+use core::{
+    fmt::Display,
+    fmt::Debug,
+    cmp::Eq,
+    cmp::PartialEq,
+    hash::Hash
+};
+
+/// Error if a value exceeds the maximum allowed value.
+#[derive(Clone, Debug, Eq, PartialEq, Hash)]
+pub struct ValueTooBigError<T: Sized + Clone + Display + Debug + Eq + PartialEq + Hash> {
+    /// Value that was disallowed.
+    pub actual: T,
+
+    /// Maximum allowed value (inclusive).
+    pub max_allowed: T,
+
+    /// Type of value.
+    pub value_type: err::ValueType,
+}
+
+
+impl<T> core::fmt::Display for ValueTooBigError<T> where
+    T: Sized + Clone + Display + Debug + Eq + PartialEq + Hash
+{
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        write!(
+            f,
+            "Error '{}' is too big to be a '{}' (maximum allowed value is '{}')",
+            self.actual,
+            self.value_type,
+            self.max_allowed
+        )
+    }
+}
+
+#[cfg(feature = "std")]
+impl<T> std::error::Error for ValueTooBigError<T> where
+    T: Sized + Clone + Display + Debug + Eq + PartialEq + Hash
+{
+    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+        None
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use std::{
+        collections::hash_map::DefaultHasher,
+        error::Error,
+        hash::{Hash, Hasher},
+        format
+    };
+
+    #[test]
+    fn fmt() {
+        assert_eq!(
+            format!("{}", ValueTooBigError{
+                actual: 3,
+                max_allowed: 2,
+                value_type: err::ValueType::IpFragmentOffset
+            }),
+            "Error '3' is too big to be a 'IP fragment offset' (maximum allowed value is '2')"
+        );
+    }
+
+    #[test]
+    fn clone_eq_hash() {
+        let err = ValueTooBigError{
+            actual: 3,
+            max_allowed: 2,
+            value_type: err::ValueType::IpFragmentOffset
+        };
+        assert_eq!(err, err.clone());
+        let hash_a = {
+            let mut hasher = DefaultHasher::new();
+            err.hash(&mut hasher);
+            hasher.finish()
+        };
+        let hash_b = {
+            let mut hasher = DefaultHasher::new();
+            err.clone().hash(&mut hasher);
+            hasher.finish()
+        };
+        assert_eq!(hash_a, hash_b);
+    }
+
+    #[cfg(feature = "std")]
+    #[test]
+    fn source() {
+        assert!(ValueTooBigError{
+            actual: 3,
+            max_allowed: 2,
+            value_type: err::ValueType::IpFragmentOffset
+        }
+        .source()
+        .is_none());
+    }
+}